The script is working quite well now.
The timetables seem to be complete (DT021 years 1-4, DT009 years 1-3). All the labels and decorations are present now. There is still a slight problem with clipping of some of the timetable entry text. Also, I have not yet verified that the information displayed is actually correct by comparing it against webtimetables, but it looks reasonable.